+++ src/zspec/zcl/definition/cluster.ts
@@ -7051,25 +7051,6 @@ const RawClusters: Readonly<Record<ClusterName, Readonly<ClusterDefinition>>> = 
             },
         },
     },
-    manuSpecificClusterAduroSmart: {
-        ID: 0xfccc,
-        attributes: {},
-        commands: {
-            cmd0: {ID: 0x00, parameters: []},
-        },
-        commandsResponse: {},
-    },
-    manuSpecificOsram: {
-        ID: 0xfc0f,
-        attributes: {},
-        commands: {
-            saveStartupParams: {ID: 0x01, parameters: []},
-            resetStartupParams: {ID: 0x02, parameters: []},
-        },
-        commandsResponse: {
-            saveStartupParamsRsp: {ID: 0x00, parameters: []},
-        },
-    },
     manuSpecificPhilips: {
         ID: 0xfc00,
         manufacturerCode: ManufacturerCode.SIGNIFY_NETHERLANDS_B_V,
@@ -7409,15 +7390,6 @@ const RawClusters: Readonly<Record<ClusterName, Readonly<ClusterDefinition>>> = 
         },
         commandsResponse: {},
     },
-    manuSpecificCentraliteHumidity: {
-        ID: 0xfc45,
-        manufacturerCode: ManufacturerCode.CENTRALITE_SYSTEMS_INC,
-        attributes: {
-            measuredValue: {ID: 0x0000, type: DataType.UINT16, write: true, max: 0xffff},
-        },
-        commands: {},
-        commandsResponse: {},
-    },
     manuSpecificSmartThingsArrivalSensor: {
         ID: 0xfc05,
         manufacturerCode: ManufacturerCode.SMARTTHINGS_INC,
@@ -7426,56 +7398,6 @@ const RawClusters: Readonly<Record<ClusterName, Readonly<ClusterDefinition>>> = 
         commandsResponse: {
             arrivalSensorNotify: {ID: 0x01, parameters: []},
         },
-    },
-    manuSpecificSamsungAccelerometer: {
-        ID: 0xfc02,
-        manufacturerCode: ManufacturerCode.SMARTTHINGS_INC,
-        attributes: {
-            motion_threshold_multiplier: {ID: 0x0000, type: DataType.UINT8, write: true, max: 0xff},
-            motion_threshold: {ID: 0x0002, type: DataType.UINT16, write: true, max: 0xffff},
-            acceleration: {ID: 0x0010, type: DataType.BITMAP8, write: true, max: 0xff},
-            x_axis: {ID: 0x0012, type: DataType.INT16, write: true, min: -32768, max: 32767},
-            y_axis: {ID: 0x0013, type: DataType.INT16, write: true, min: -32768, max: 32767},
-            z_axis: {ID: 0x0014, type: DataType.INT16, write: true, min: -32768, max: 32767},
-        },
-        commands: {},
-        commandsResponse: {},
-    },
-    sprutVoc: {
-        ID: 0x6601,
-        manufacturerCode: 26214,
-        attributes: {
-            voc: {ID: 0x6600, type: DataType.UINT16, write: true, max: 0xffff},
-        },
-        commands: {},
-        commandsResponse: {},
-    },
-    sprutNoise: {
-        ID: 0x6602,
-        manufacturerCode: 26214,
-        attributes: {
-            noise: {ID: 0x6600, type: DataType.SINGLE_PREC, write: true},
-            noiseDetected: {ID: 0x6601, type: DataType.BITMAP8, write: true},
-            noiseDetectLevel: {ID: 0x6602, type: DataType.SINGLE_PREC, write: true},
-            noiseAfterDetectDelay: {ID: 0x6603, type: DataType.UINT16, write: true, max: 0xffff},
-        },
-        commands: {},
-        commandsResponse: {},
-    },
-    sprutIrBlaster: {
-        ID: 0x6603,
-        manufacturerCode: 26214,
-        attributes: {},
-        commands: {
-            playStore: {ID: 0x00, parameters: [{name: "param", type: DataType.UINT8, max: 0xff}]},
-            learnStart: {ID: 0x01, parameters: [{name: "value", type: DataType.UINT8, max: 0xff}]},
-            learnStop: {ID: 0x02, parameters: [{name: "value", type: DataType.UINT8, max: 0xff}]},
-            clearStore: {ID: 0x03, parameters: []},
-            playRam: {ID: 0x04, parameters: []},
-            learnRamStart: {ID: 0x05, parameters: []},
-            learnRamStop: {ID: 0x06, parameters: []},
-        },
-        commandsResponse: {},
     },
     manuSpecificSiglisZigfred: {
         ID: 0xfc42,